+++ glide-core/redis-rs/redis/src/cluster_async/mod.rs
@@ -338,6 +338,15 @@ where
             .await
     }
 
+    /// Update the name used for all cluster connections
+    pub async fn update_connection_client_name(
+        &mut self,
+        client_name: Option<String>,
+    ) -> RedisResult<Value> {
+        self.route_operation_request(Operation::UpdateConnectionClientName(client_name))
+            .await
+    }
+
     /// Get the username used to authenticate with all cluster servers
     pub async fn get_username(&mut self) -> RedisResult<Value> {
         self.route_operation_request(Operation::GetUsername).await
@@ -662,6 +671,7 @@ enum CmdArg<C> {
 enum Operation {
     UpdateConnectionPassword(Option<String>),
     UpdateConnectionDatabase(i64),
+    UpdateConnectionClientName(Option<String>),
     GetUsername,
 }
 
@@ -2601,6 +2611,11 @@ where
                 }
                 Operation::UpdateConnectionDatabase(database_id) => {
                     core.set_cluster_param(|params| params.database_id = database_id)
+                        .expect(MUTEX_WRITE_ERR);
+                    Ok(Response::Single(Value::Okay))
+                }
+                Operation::UpdateConnectionClientName(client_name) => {
+                    core.set_cluster_param(|params| params.client_name = client_name)
                         .expect(MUTEX_WRITE_ERR);
                     Ok(Response::Single(Value::Okay))
                 }